Refactor, move function to sdk
