+++ mtdaws/sdk.go
@@ -1,4 +1,4 @@
-package mtd_aws
+package mtdaws
 
 import (
 	"context"
@@ -9,8 +9,12 @@ import (
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
+	"github.com/google/uuid"
+	"github.com/thefeli73/polemos/state"
 )
-func New_config(region string, credentials string) aws.Config {
+
+// NewConfig creates a AWS config for a specific region
+func NewConfig(region string, credentials string) aws.Config {
 	// Create a new AWS config
 	cfg, err := config.LoadDefaultConfig(context.TODO(), config.WithSharedConfigFiles([]string{credentials}), config.WithRegion(region))
 	if err != nil {
@@ -20,7 +24,28 @@ func New_config(region string, credentials string) aws.Config {
 	}
 	return cfg
 }
-func Instance_info(config aws.Config, instanceID string) {
+
+// IndexInstances scans all configured regions for instances and add them to services
+func IndexInstances(config state.Config) {
+	for _, region := range config.AWS.Regions {
+		//fmt.Println("Listing instances in region:", region)
+		awsConfig := NewConfig(region, config.AWS.CredentialsPath)
+		instances, err := Instances(awsConfig)
+		if err != nil {
+			fmt.Println("Error listing instances:", err)
+			continue
+		}
+
+		for _, instance := range instances {
+			InstanceInfo(awsConfig, *instance.InstanceId)
+			u := uuid.New()
+			_ = u
+		}
+	}
+}
+
+// InstanceInfo collects info about a specific instance in a region
+func InstanceInfo(config aws.Config, instanceID string) {
 	// Create a new EC2 service client
 	svc := ec2.NewFromConfig(config)
 
@@ -45,6 +70,7 @@ func Instance_info(config aws.Config, instanceID string) {
 	fmt.Println("Private IP Address:", aws.ToString(instance.PrivateIpAddress))
 }
 
+// Instances returns all instances for a config i.e. a region
 func Instances(config aws.Config) ([]*types.Instance, error) {
 	svc := ec2.NewFromConfig(config)
